pg/catalog: keep the metadata loader in one file, read top down
Merge metadata_load_ast.go into metadata_load.go, ordered as a load
runs: the API, the objects, collecting, ordering, installing, statement
builders, stand-ins, and text parsing. Behavior does not change, apart
from the order of enums against composite types where nothing orders
them.

- The kinds are declared in install order, so objects sort by a
  comparison instead of a string key built from a table of labels.
- installGroup holds a cycle's passes, so LoadMetadata reads as
  collect, order, install.
- installStandIn installs a stand-in directly instead of returning a
  closure, and replaceWithStandIn records the outcome.
- Partitions are walked by one forEachPartition, and their constraints
  are collected with the other Full-only objects.
- The tests move into one file, and the large dependency fixture splits
  into tests for tables, view bodies, and overloaded calls.
